+++ src/Ampersand/Output/ToJSON/JSONutils.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 module Ampersand.Output.ToJSON.JSONutils 
-  (writeJSONFile, JSON(..), JSON'(..), ToJSON(..)
+  (JSON(..), JSON'(..), ToJSON(..)
   , module Ampersand.Basics
   , module Ampersand.Classes
   , module Ampersand.Core.ParseTree
@@ -23,23 +23,11 @@ import           Ampersand.FSpec.ToFSpec.Populated
 import           Ampersand.FSpec.FSpec
 import           Ampersand.FSpec.SQL (sqlQuery,sqlQueryWithPlaceholder,placeHolderSQL,broadQueryWithPlaceholder) 
 import           Ampersand.Misc.HasClasses
-import           Ampersand.Prototype.ProtoUtil(getGenericsDir)
 import           Data.Aeson hiding (Options)
-import qualified Data.Aeson.Types as AT 
-import           Data.Aeson.Encode.Pretty
-import qualified RIO.ByteString.Lazy as BL
+import qualified Data.Aeson.Types as AT
 import qualified RIO.List as L
 import qualified RIO.Text as T
 import           GHC.Generics
-import           System.FilePath
-import           System.Directory
-
-writeJSONFile :: (ToJSON a, HasLogFunc env) => 
-                 FilePath -> a -> RIO env ()
-writeJSONFile fullFile x = do
-    logDebug $ "  Generating "<>display (T.pack fullFile) 
-    liftIO $ createDirectoryIfMissing True (takeDirectory fullFile)
-    liftIO $ BL.writeFile fullFile (encodePretty x)
 
 -- | We use aeson to generate .json in a simple and efficient way.
 --   For details, see http://hackage.haskell.org/package/aeson/docs/Data-Aeson.html#t:ToJSON
@@ -50,7 +38,7 @@ class (GToJSON Zero (Rep b), Generic b) => JSON a b | b -> a where
   amp2Jason = genericToJSON ampersandDefault
 -- | Same as JSON, but different constraints for fromAmpersand'
 class (GToJSON Zero (Rep b), Generic b) => JSON' a b | b -> a where
-  fromAmpersand' :: (HasProtoOpts env, Show env) 
+  fromAmpersand' :: (Show env, HasProtoOpts env) 
        => env -> FSpec -> a -> b
   amp2Jason' :: b -> Value
   amp2Jason' = genericToJSON ampersandDefault
@@ -64,9 +52,10 @@ ampersandDefault = defaultOptions {AT.fieldLabelModifier = alterLabel}
     -- the .json side. In our case, we strip all characters upto the first occurence
     -- of the prefix "JSON" (which is mandatory). in the rest of that string, we 
     -- substitute all underscores with dots.
+    alterLabel :: String -> String
     alterLabel str =
       case filter (L.isPrefixOf pfx) (L.tails str) of
-        []  -> fatal ("Label at Haskell side must contain `JSON`: "++str)
+        []  -> fatal ("Label at Haskell side must contain `JSON`: "<>T.pack str)
         h:_ -> replace '_' '.' . snd . L.splitAt (length pfx) $ h
       where pfx = "JSON"
 
@@ -77,5 +66,3 @@ replace :: Eq a =>
         -> [a] -- ^ Input list
         -> [a] -- ^ Output list
 replace x y = map (\z -> if z == x then y else z)
-  
-